Fixed Service layer functionalities being in bean layer (controllers)

+++ src/main/java/danielrichtersz/services/impl/VoteServiceImpl.java
@@ -7,6 +7,7 @@ import danielrichtersz.models.Vote;
 import danielrichtersz.models.enums.TypeVote;
 import danielrichtersz.repositories.interfaces.CommentRepository;
 import danielrichtersz.repositories.interfaces.PostRepository;
+import danielrichtersz.repositories.interfaces.RedditorRepository;
 import danielrichtersz.repositories.interfaces.VoteRepository;
 import danielrichtersz.services.interfaces.VoteService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -24,40 +25,36 @@ public class VoteServiceImpl implements VoteService {
     @Autowired
     private VoteRepository voteRepository;
 
+    @Autowired
+    private RedditorRepository redditorRepository;
+
     @Override
-    public Vote createVote(Vote vote) {
+    public Vote updateOrCreateVote(Long postId, String username, TypeVote typeVote) {
+        Post post = postRepository.findById(postId).get();
+        Redditor redditor = redditorRepository.findByUsername(username);
+        Vote vote = voteRepository.findByOwnerAndPost(redditor, post);
+        if (vote == null) {
+            return voteRepository.save(new Vote(post, redditor, typeVote));
+        }
+        vote.setTypeVote(typeVote);
         return voteRepository.save(vote);
     }
 
     @Override
-    public Vote updateVote(Vote vote) {
-        return voteRepository.save(vote);
+    public void deleteVote(Long postId, String username) {
+        Post post = postRepository.findById(postId).get();
+        Redditor redditor = redditorRepository.findByUsername(username);
+        Vote vote = voteRepository.findByOwnerAndPost(redditor, post);
+
+        if (vote != null) {
+            voteRepository.delete(vote);
+        }
     }
 
     @Override
-    public Post upvotePost(Post post, Redditor redditor, TypeVote vote) {
-        post.addVote(new Vote(post, redditor, vote));
-        return postRepository.save(post);
-    }
-
-    @Override
-    public Comment upvoteComment(Comment comment, Redditor redditor, TypeVote vote) {
-        comment.addVote(new Vote(comment, redditor, vote));
-        return commentRepository.save(comment);
-    }
-
-    @Override
-    public Vote getVote(Post post, Redditor owner) {
-        return voteRepository.findByOwnerAndPost(owner, post);
-    }
-
-    @Override
-    public Vote getVote(Comment comment, Redditor owner) {
-        return voteRepository.findByOwnerAndComment(owner, comment);
-    }
-
-    @Override
-    public void deleteVote(Vote vote) {
-        voteRepository.delete(vote);
+    public Vote getVote(Long postId, String username) {
+        Post post = postRepository.findById(postId).get();
+        Redditor redditor = redditorRepository.findByUsername(username);
+        return voteRepository.findByOwnerAndPost(redditor, post);
     }
 }
